+++ CI/src/integration_tests_for_virtual_cluster/test_ssh_hook.py
@@ -70,7 +70,7 @@ class TestSshHook(unittest.TestCase):
                     "^com.hooks.ssh.enabled$": "^true$"
                 }
             },
-            "stages": ["createRuntime"]
+            "stages": ["createRuntime", "poststop"]
         }
         return hook_config