+++ src/hooks/stdout_stderr_test/main.cpp
@@ -18,8 +18,8 @@ using namespace sarus;
 int main(int argc, char* argv[]) {
     boost::filesystem::path bundleDir;
     pid_t pidOfContainer;
-    std::tie(bundleDir, pidOfContainer) = common::hook::parseStateOfContainerFromStdin();
-    auto json = sarus::common::readJSON(bundleDir / "config.json");
+    auto containerState = common::hook::parseStateOfContainerFromStdin();
+    auto json = sarus::common::readJSON(containerState.bundle() / "config.json");
     common::hook::applyLoggingConfigIfAvailable(json);
 
     std::cout << "hook's stdout" << std::endl;